x/nftmngr/types/messages_schema_attribute_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/thesixnetwork/sixnft/testutil/sample"
)

func TestMsgCreateSchemaAttribute_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateSchemaAttribute
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateSchemaAttribute{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateSchemaAttribute{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateSchemaAttribute_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateSchemaAttribute
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateSchemaAttribute{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateSchemaAttribute{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteSchemaAttribute_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteSchemaAttribute
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteSchemaAttribute{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteSchemaAttribute{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
